preparing to move from foreignkey to m2m

+++ trashit/docker-entrypoint.sh
@@ -2,14 +2,14 @@
 # docker-entrypoint.sh
 
 # If this is going to be a cron container, set up the crontab.
-# if [ "$1" = cron ]; then
-#   if test -f "$FILE"; then
+if [ "$1" = cron ]; then
+  if test -f "$FILE"; then
     echo 'Start removed manage added prod'
     rm manage.py && mv manage.production.py manage.py
     rm trashit/wsgi.py && mv trashit/wsgi.production.py trashit/wsgi.py
-#   fi
+  fi
 
-# fi
+fi
 python manage.py makemigrations --noinput
 python manage.py migrate --noinput
 python manage.py collectstatic --noinput
