+++ trashit/start.sh
@@ -1,9 +1,9 @@
 FILE=manage.production.py
-# if test -f "$FILE"; then
+if test -f "$FILE"; then
 echo 'Start removed manage added prod'
 rm manage.py && mv manage.production.py manage.py
 rm trashit/wsgi.py && mv trashit/wsgi.production.py trashit/wsgi.py
-# fi
+fi
 python manage.py makemigrations --noinput
 python manage.py migrate --noinput
 python manage.py collectstatic --noinput
